+++ CHANGES.rst
@@ -9,6 +9,8 @@ Version 0.14
 - Added support for edge as browser.
 - Added support for platforms that lack ``SpooledTemporaryFile``.
 - Add support for etag handling through if-match
+- Added support for the SameSite cookie attribute.
+
 
 Version 0.13
 ------------